Update statkeeper label displays on entry page
Create list of Statkeepers, displaying the statkeepers name, along with
an icon based on whether it's a player, team, or league statkeeper.

+++ lib/model/player_stats.dart
@@ -1,7 +1,7 @@
 import 'package:meta/meta.dart';
 
 class PlayerStats {
-  BigInt id;
+  double id;
   String firestoreID;
   String teamFirestoreID;
   String name;
